+++ src/data/reduceable.h
@@ -10,10 +10,10 @@
 class Reduceable : public Serializable {
 
 public:
-    virtual std::vector<int> serialize() const override = 0;
-    virtual void deserialize(const std::vector<int>& packed) override = 0;
+    virtual std::vector<uint8_t> serialize() const override = 0;
+    virtual void deserialize(const std::vector<uint8_t>& packed) override = 0;
     virtual void merge(const Reduceable& other) = 0;
-    virtual std::unique_ptr<Reduceable> getDeserialized(const std::vector<int>& packed) const = 0;
+    virtual std::unique_ptr<Reduceable> getDeserialized(const std::vector<uint8_t>& packed) const = 0;
     virtual bool isEmpty() = 0;
 
     std::set<int> allReduce(MPI_Comm& comm);
