+++ src/data/serializable.h
@@ -3,12 +3,13 @@
 #define DOMPASCH_CUCKOO_REBALANCER_SERIALIZABLE
 
 #include <vector>
+#include <cstdint>
 
 class Serializable {
 
 public:
-    virtual std::vector<int> serialize() const = 0;
-    virtual void deserialize(const std::vector<int>& packed) = 0;
+    virtual std::vector<uint8_t> serialize() const = 0;
+    virtual void deserialize(const std::vector<uint8_t>& packed) = 0;
 };
 
 #endif
